pyproject.toml
[project]
name = "video-processor"
version = "0.1.0"
description = "A video processing project using Video-LLaVA"
requires-python = ">=3.11.2,<3.12"
dependencies = [
    "torch==2.2.1+cu118",
    "torchvision==0.17.1+cu118",
    "torchaudio==2.2.1+cu118",
    "transformers==4.42.4",
    "flash-attn==2.2.0",
    "sentence-transformers==2.7.0",
    "av",
    "qdrant-client>=1.7.0",
    "tqdm",
    "ffmpeg-python",
    "python-dotenv",
    "openai>=1.12.0",
    "numpy>=1.26.4"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
